Move trading_session fixture to a new conftest
